requirements.txt
aiofiles==22.1.0
fastapi==0.88.0
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.8.3
Pillow==9.3.0
psycopg2==2.9.5
pydantic==1.10.4
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
passlib
bcrypt